+++ ChangeLog.md
@@ -15,7 +15,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added 
 
-- ...
+- jsonl/spitln function